add MD25. delete unnecessary files. replace DEBUG by DEBUG_UDP. replace all int64_t by int32_t.

+++ asservissement/asserv/codeurs/Codeur.cpp
@@ -21,7 +21,7 @@ void Codeur::reset(void)
 }
 
 
-int Codeur::getCount(void)
+long Codeur::getCount(void)
 {
     return count;
 }
